Cleanup (#310)
* remove redundant UAP
since the lib already support netstandard1.3  UAP is supproted OOTB via the net standard compat matrix https://docs.microsoft.com/en-us/dotnet/standard/net-standard
* redundant else
* redundant private set
* use expression bodies for simple properties
* redundant usings
* spelling
* remove dodgy casting
* redundant casts
* redundant constructors
* redundant sets
* use built in type
* use expression body
* remove redundant async state machines
* use var where evident
* fix spelling
* use pattern matching
* use build props file
* remove redundant type specifications
* move signing to props

+++ src/Microsoft.Azure.ServiceBus/Primitives/ConcurrentRandom.cs
@@ -27,7 +27,7 @@ namespace Microsoft.Azure.ServiceBus.Primitives
             var buffer = new byte[8];
             GetThreadLocalRandom().NextBytes(buffer);
             var ulongValue = (long)BitConverter.ToUInt64(buffer, 0);
-            return Math.Abs((long)ulongValue);
+            return Math.Abs(ulongValue);
         }
 
         static Random GetThreadLocalRandom()
